Reorganize project in separate view controllers
This patch contains a lot of changes touching the entire codebase. In
dependence of https://goo.gl/KLryQJ and the whole series of articles,
the storyboard is removed and additional, small view controllers are
added to the project.

For now, the `MessageViewController` shows only the name and the
official details of the occured error case within it's labels.